Add files via upload
This .py file  code of Python-based dice game for 2 to 4 players. Players take turns rolling a die (1-6) to accumulate points, aiming to reach or exceed a score of 50. If a player rolls a 1 on their first roll, their turn ends with zero points. If they roll a 1 later, their turn ends, and the next player starts. At the end, the player with the highest score wins.

This is my first attempt at learning Python. If you find any errors or oversights, please feel free to drop me a line!"